+++ dialect.go
@@ -1,16 +1,9 @@
 package dbr
 
-import "time"
+import (
+	"github.com/gocraft/dbr/v2/dialect"
+)
 
 // Dialect abstracts database driver differences in encoding
 // types, and placeholders.
-type Dialect interface {
-	QuoteIdent(id string) string
-
-	EncodeString(s string) string
-	EncodeBool(b bool) string
-	EncodeTime(t time.Time) string
-	EncodeBytes(b []byte) string
-
-	Placeholder(n int) string
-}
+type Dialect = dialect.Dialect
